requirements.txt
# Web framework
fastapi==0.100.0
uvicorn==0.23.2
pydantic==1.10.13

# Machine Learning
scikit-learn==1.3.0
numpy==1.26.4
pandas==1.5.3
joblib==1.3.2

# OpenTelemetry Tracing
# Using a version set compatible with Python 3.11
opentelemetry-api==1.9.0
opentelemetry-sdk==1.9.0
opentelemetry-exporter-gcp-trace==1.9.0
